Move HandlePanics function to its own package
